Move new exception from fi.nls -> org.oskari
