+++ service-search-opendata/src/main/java/fi/nls/oskari/search/What3WordsSearchChannel.java
@@ -9,8 +9,7 @@ import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.map.geometry.ProjectionHelper;
 import fi.nls.oskari.search.channel.SearchChannel;
-import fi.nls.oskari.service.ComponentSkippedRuntimeException;
-import fi.nls.oskari.util.ConversionHelper;
+import org.oskari.component.ComponentSkippedRuntimeException;
 import fi.nls.oskari.util.IOHelper;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
